backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pymongo>=4.5.0
python-dotenv>=1.0.0
resend>=0.7.0
cloudinary>=1.36.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
scikit-learn>=1.3.0
timm>=0.9.0
albumentations>=1.3.0

